requirements.txt
numpy
opencv-python
matplotlib
scikit-learn
streamlit
pillow